+++ packages/nimble-components/src/button/tests/button.spec.ts
@@ -8,12 +8,8 @@ describe('Button', () => {
         return await fixture<Button>(html`<${buttonTag}></${buttonTag}>`);
     }
 
-    it('should export its tag', () => {
-        expect(buttonTag).toBe('nimble-button');
-    });
-
     it('can construct an element instance', () => {
-        expect(document.createElement('nimble-button')).toBeInstanceOf(Button);
+        expect(document.createElement(buttonTag)).toBeInstanceOf(Button);
     });
 
     it('should default tabIndex on the internal button to 0', async () => {