+++ packages/nimble-components/src/menu-item/tests/menu-item.spec.ts
@@ -1,13 +1,7 @@
 import { MenuItem, menuItemTag } from '..';
 
 describe('MenuItem', () => {
-    it('should export its tag', () => {
-        expect(menuItemTag).toBe('nimble-menu-item');
-    });
-
     it('can construct an element instance', () => {
-        expect(document.createElement('nimble-menu-item')).toBeInstanceOf(
-            MenuItem
-        );
+        expect(document.createElement(menuItemTag)).toBeInstanceOf(MenuItem);
     });
 });