+++ packages/nimble-components/src/radio-group/tests/radio-group.spec.ts
@@ -1,12 +1,8 @@
 import { RadioGroup, radioGroupTag } from '..';
 
 describe('Radio Group', () => {
-    it('should export its tag', () => {
-        expect(radioGroupTag).toBe('nimble-radio-group');
-    });
-
     it('can construct an element instance', () => {
-        expect(document.createElement('nimble-radio-group')).toBeInstanceOf(
+        expect(document.createElement(radioGroupTag)).toBeInstanceOf(
             RadioGroup
         );
     });