+++ packages/nimble-components/src/radio/tests/radio.spec.ts
@@ -1,11 +1,7 @@
 import { Radio, radioTag } from '..';
 
 describe('Radio', () => {
-    it('should export its tag', () => {
-        expect(radioTag).toBe('nimble-radio');
-    });
-
     it('can construct an element instance', () => {
-        expect(document.createElement('nimble-radio')).toBeInstanceOf(Radio);
+        expect(document.createElement(radioTag)).toBeInstanceOf(Radio);
     });
 });