+++ packages/nimble-components/src/tree-item/tests/tree-item.spec.ts
@@ -1,13 +1,7 @@
 import { TreeItem, treeItemTag } from '..';
 
 describe('TreeItem', () => {
-    it('should export its tag', () => {
-        expect(treeItemTag).toBe('nimble-tree-item');
-    });
-
     it('can construct an element instance', () => {
-        expect(document.createElement('nimble-tree-item')).toBeInstanceOf(
-            TreeItem
-        );
+        expect(document.createElement(treeItemTag)).toBeInstanceOf(TreeItem);
     });
 });