+++ packages/storybook/src/nimble/patterns/tabs/types.ts
@@ -1,0 +1,7 @@
+export const ExampleTabsType = {
+    simpleTabs: 'SimpleTabs',
+    manyTabs: 'ManyTabs',
+    wideTabs: 'WideTabs'
+} as const;
+export type ExampleTabsType =
+    (typeof ExampleTabsType)[keyof typeof ExampleTabsType];