+++ packages/storybook/src/nimble/tabs/tabs-matrix.stories.ts
@@ -28,12 +28,17 @@ const metadata: Meta = {
 
 export default metadata;
 
+const widthStates = ['', '250px'] as const;
+type WidthState = (typeof widthStates)[number];
+
 // prettier-ignore
 const component = (
     toolbar: TabsToolbarState,
-    [disabledName, disabled]: DisabledState
+    [disabledName, disabled]: DisabledState,
+    widthValue: WidthState
+
 ): ViewTemplate => html`
-    <${tabsTag} style="padding: 15px;">
+    <${tabsTag} style="padding: 15px;${widthValue ? ` width: ${widthValue};` : ''}">
         ${when(() => toolbar, html`
             <${tabsToolbarTag}>
                 <${buttonTag} appearance="ghost">Toolbar Button</${buttonTag}>
@@ -45,14 +50,16 @@ const component = (
             Tab Two ${() => disabledName}
         </${tabTag}>
         <${tabTag} hidden>Tab Three</${tabTag}>
+        <${tabTag}>Tab Four</${tabTag}>
         <${tabPanelTag}>Tab content one</${tabPanelTag}>
         <${tabPanelTag}>Tab content two</${tabPanelTag}>
         <${tabPanelTag}>Tab content three</${tabPanelTag}>
+        <${tabPanelTag}>Tab content four</${tabPanelTag}>
     </${tabsTag}>
 `;
 
 export const tabsThemeMatrix: StoryFn = createMatrixThemeStory(
-    createMatrix(component, [tabsToolbarStates, disabledStates])
+    createMatrix(component, [tabsToolbarStates, disabledStates, widthStates])
 );
 
 export const hiddenTabs: StoryFn = createStory(
@@ -77,8 +84,8 @@ export const textCustomized: StoryFn = createMatrixThemeStory(
 );
 
 export const panelOverflow: StoryFn = createStory(html`
-    <nimble-tabs style="height: 120px; width: 400px;">
-        <nimble-tab>Tab One</nimble-tab>
-        <nimble-tab-panel style="width: 450px;">${loremIpsum}</nimble-tab-panel>
-    </nimble-tabs>
+    <${tabsTag} style="height: 120px; width: 400px;">
+        <${tabTag}>Tab One</${tabTag}>
+        <${tabPanelTag} style="width: 450px;">${loremIpsum}</${tabPanelTag}>
+    </${tabsTag}>
 `);